iox-#1036 Extract builder pattern and move it into separate design pattern header
